@@ -80,7 +80,10 @@ class PhpSourceHelperCss
         {
           if (preg_match("/(->|::)($method)(\\()/", $line))
           {
-            $this->task->logError("Unexpected usage of method '%s' at line %s:%d", $method, $source['src_path'], $i + 1);
+            $this->task->logError("Unexpected usage of method '%s' at line %s:%d",
+                                  $method,
+                                  $source['src_path'],
+                                  $i + 1);
           }
         }
       }
@@ -107,12 +110,12 @@ class PhpSourceHelperCss
   {
     [$path, $expression] = $this->deriveResourcePath($matches, $qualifiedName, $imports, $namespace);
 
-    $this->task->logVerbose('    found %s (%s:%d)', Path::makeRelative($path, $this->buildPath), $expression, $lineno);
+    $this->task->logVerbose('    found %s (%s:%d).', Path::makeRelative($path, $this->buildPath), $expression, $lineno);
 
     $resource = $this->store->resourceSearchByPath($path);
     if ($resource===null)
     {
-      $this->task->logError("Unable to find resource '%s' found at %s:%d", $path, $source['src_path'], $lineno);
+      $this->task->logError("Unable to find resource '%s' found at %s:%d.", $path, $source['src_path'], $lineno);
 
       return;
     }
@@ -193,7 +196,7 @@ class PhpSourceHelperCss
    */
   private function resolveFullPathOfResource(string $resourceName): string
   {
-    if (substr($resourceName, 0, 1)==='/')
+    if (str_starts_with($resourceName, '/'))
     {
       $fullPath = Path::join($this->parentResourcePath, $resourceName);
     }
